third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['text-decoration'] = "none" should set the property value
PASS e.style['text-decoration'] = "line-through" should set the property value
FAIL e.style['text-decoration'] = "solid" should set the property value assert_equals: serialization should be canonical expected "none" but got "solid"
FAIL e.style['text-decoration'] = "currentcolor" should set the property value assert_equals: serialization should be canonical expected "none" but got "currentcolor"
FAIL e.style['text-decoration'] = "auto" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['text-decoration'] = "from-font" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['text-decoration'] = "10px" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['text-decoration'] = "double overline underline" should set the property value
PASS e.style['text-decoration'] = "underline overline line-through red" should set the property value
PASS e.style['text-decoration'] = "rgba(10, 20, 30, 0.4) dotted" should set the property value
FAIL e.style['text-decoration'] = "overline green from-font" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['text-decoration'] = "underline dashed green" should set the property value
FAIL e.style['text-decoration'] = "underline auto" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['text-decoration'] = "line-through 20px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['text-decoration'] = "overline 3em" should set the property value assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
